Create README.md
This folder includes writeups for challenges solved on Hack The Box (HTB), a platform for intermediate to advanced penetration testing and CTF competitions. It covers a wide range of topics like web exploitation, privilege escalation, and reverse engineering. The writeups here highlight the skills and techniques used to exploit real-world security vulnerabilities.